Fix media playback issue when audio sound device is not available
Video freezes when played for the first time after being pushed
to the DUT.

Media playback is freezing as opening of a sound card is failing
due to board not connected.

Introduced a counter to check the sound card unavailability failures.
When the counter reaches maximum count, fallback to dummy sound card.

Tests Done:
- Verified Audio playback with external input audio clk
- Verified Audio playback without external input audio clk
- Verified Video playback with external input audio clk
- Verified Video playback without external input audio clk
- Verified Audio capture with external input audio clk
- Verified Audio capture without external input audio clk

Tracked-On: OAM-112568
Signed-off-by: Deepa K G <g.k.deepa@intel.com>
Signed-off-by: Martin, Chen <haochuan.z.chen@intel.com>