Enable error logging in camera config
we need to bind v4l-subdev to v4l2 video device to poll the soib link status

when the sensor link is broken in short time, we should control v4l-subdev
the binding is set in the ivi_camera_config.xml file.

Tests done: Tested with xml file getting updated on the target device.

Tracked-On: OAM-114095
Signed-off-by: shivasku82 <shiva.kumara.rudrappa@intel.com>